refactor(vpc): move vpc resources to service path
